pyproject.toml
[project]
name = "prefect-examples"
version = "0.1.0"
description = "Progressive Prefect 3 examples — from hello-world to production patterns."
requires-python = ">=3.13"
dependencies = [
    "prefect>=3.6",
    "prefect-aws",
    "prefect-dhis2",
    "pandas",
    "pyarrow",
    "python-dotenv",
    "geopandas",
    "s3fs",
]

[dependency-groups]
dev = [
    "mkdocs<2",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mypy>=1",
    "pandas-stubs>=3.0.0.260204",
    "pytest>=8",
    "ruff>=0.9",
]

[build-system]
requires = ["uv_build>=0.7"]
build-backend = "uv_build"

[tool.uv.sources]
prefect-dhis2 = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]

# ---------- Ruff ----------
[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.pydocstyle]
convention = "google"

# ---------- mypy ----------
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "prefect.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["geopandas", "geopandas.*", "shapely", "shapely.*"]
ignore_missing_imports = true

# ---------- pytest ----------
[tool.pytest.ini_options]
testpaths = ["tests"]
